requirements.txt
Django
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.1.0
boto3
geopy==2.4.1
llama-index>=0.10.0
openai>=1.0.0
aiohttp>=3.8.0
typing-extensions>=4.5.0
pydantic>=2.0.0
